requirements.txt
langchain_community 
beautifulsoup4
html2text
python-dotenv
google-generativeai
langchain-google-genai
langchain
faiss-cpu
tiktoken
gradio